refactor proto dirs
