+++ pkg/testing/client.go
@@ -9,7 +9,7 @@ import (
 	gethtypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/hyperledger-labs/yui-ibc-solidity/pkg/chains"
 	"github.com/hyperledger-labs/yui-ibc-solidity/pkg/client"
-	ibcclient "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/client"
+	ibcclient "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/core/client"
 )
 
 type LightClient struct {
