+++ proto/apps/20-transfer/FungibleTokenPacketData.proto
@@ -1,6 +1,6 @@
 syntax = "proto3";
 
-option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/app";
+option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/apps/transfer";
 
 import "solidity-protobuf-extensions.proto";
 option (.solidity.file_options) = { location: "@hyperledger-labs/yui-ibc-solidity/contracts/proto" };
