+++ proto/clients/ibft2/IBFT2.proto
@@ -4,9 +4,9 @@ package ibc.lightclients.ibft2.v1;
 
 import "gogoproto/gogo.proto";
 import "solidity-protobuf-extensions.proto";
-import "client/Client.proto";
+import "core/02-client/Client.proto";
 
-option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/client/ibft2";
+option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/clients/ibft2";
 option (gogoproto.goproto_getters_all)  = false;
 option (.solidity.file_options) = { location: "@hyperledger-labs/yui-ibc-solidity/contracts/proto" };
