+++ proto/clients/mock/MockClient.proto
@@ -4,9 +4,9 @@ package ibc.lightclients.mock.v1;
 
 import "gogoproto/gogo.proto";
 import "solidity-protobuf-extensions.proto";
-import "client/Client.proto";
+import "core/02-client/Client.proto";
 
-option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/client/mock";
+option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/clients/mock";
 option (gogoproto.goproto_getters_all)  = false;
 option (.solidity.file_options) = { location: "@hyperledger-labs/yui-ibc-solidity/contracts/proto" };
