+++ proto/core/04-channel/Channel.proto
@@ -1,12 +1,12 @@
 syntax = "proto3";
 
-option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/channel";
+option go_package = "github.com/hyperledger-labs/yui-ibc-solidity/pkg/ibc/core/channel";
 
 import "solidity-protobuf-extensions.proto";
 option (.solidity.file_options) = { location: "@hyperledger-labs/yui-ibc-solidity/contracts/proto" };
 
 import "gogoproto/gogo.proto";
-import "client/Client.proto";
+import "core/02-client/Client.proto";
 
 // Channel defines pipeline for exactly-once packet delivery between specific
 // modules on separate blockchains, which has at least one end capable of
